requirements.txt
# Sovereign Computer — Python dependencies
# Install: pip install -r requirements.txt

langgraph>=0.1.0
langchain>=0.2.0
langchain-ollama>=0.1.0
langchain-community>=0.2.0
crewai>=0.28.0
chromadb>=0.5.0
playwright>=1.44.0
docker>=7.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
